# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.UserInput:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  handlers.WalletInput:
    properties:
      name:
        type: string
      type:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@expense-tracker.local
    name: API Support
    url: http://www.swagger.io/support
  description: Expense Tracker REST API for users, categories, transactions, balances,
    wallets.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /api/users:
    get:
      description: Get all users (protected)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a user (protected)
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Create user
      tags:
      - Users
  /api/users/{id}:
    delete:
      description: Delete user by id (protected)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Delete user
      tags:
      - Users
    get:
      description: Get user by id (protected)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user info (protected)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Update user
      tags:
      - Users
  /api/wallets:
    get:
      description: Get all wallets for a user (protected)
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get wallets
      tags:
      - Wallets
    post:
      consumes:
      - application/json
      description: Create wallet for a user (protected)
      parameters:
      - description: Create wallet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.WalletInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Create wallet
      tags:
      - Wallets
  /api/wallets/{id}:
    delete:
      description: Delete a wallet by id for a user (protected)
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Delete wallet
      tags:
      - Wallets
    get:
      description: Get wallet by id for a user (protected)
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get wallet
      tags:
      - Wallets
    put:
      consumes:
      - application/json
      description: Update wallet details (protected)
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update wallet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.WalletInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Update wallet
      tags:
      - Wallets
schemes:
- http
swagger: "2.0"